ci: lint our security-insights.yml in CI workflow
This adds a Makefile task named `lintinsights` that
downloads the v2.1.0 Insights schema and then
runs `cue vet security-insights.yml schema.cue` to
ensure that we maintain a valid insights file.

The CI workflow is updated to run this new make
task, resulting in checks failing when a PR changes
the `security-insights.yml` file causing it to no
longer conform to the schema.

Signed-off-by: Travis Truman <trumant@gmail.com>